refactor: adapter삭제, pub/sub 모듈로 분리
