--- rento-adapter/Dockerfile
@@ -1,4 +1,0 @@
-FROM eclipse-temurin:17-jdk-alpine
-COPY build/libs/*.jar app.jar
-
-ENTRYPOINT ["java", "-jar", "app.jar"]
