--- rento-adapter/build.gradle
@@ -1,65 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot'
-    id 'io.spring.dependency-management'
-}
-
-group = 'com.kbe5'
-version = '1.0-SNAPSHOT'
-
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
-dependencies {
-    implementation project(':common')
-    implementation project(':domain')
-    implementation project(':infra')
-
-    // Swagger OpenAPI 3
-    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.7.0'
-
-    implementation 'io.github.cdimascio:java-dotenv:5.2.2'
-    implementation 'org.springframework.security:spring-security-crypto'
-    implementation 'org.springframework.boot:spring-boot-starter-validation:3.4.3'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.3'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.3'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.3'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-
-    compileOnly 'org.projectlombok:lombok'
-
-    // Security
-//    implementation 'org.springframework.boot:spring-boot-starter-security'
-
-    // https://mvnrepository.com/artifact/org.springframework/spring-jdbc
-    implementation 'org.springframework:spring-jdbc:6.2.7'
-
-    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-amqp
-    implementation 'org.springframework.boot:spring-boot-starter-amqp:3.5.0'
-
-    runtimeOnly 'com.h2database:h2'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-
-    annotationProcessor 'org.projectlombok:lombok'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    testImplementation 'org.assertj:assertj-core:3.24.2'
-}
-
-test {
-    useJUnitPlatform()
-}
-
-bootJar {
-    enabled = true
-}
-
-jar {
-    enabled = false
-}
