--- rento-adapter/src/main/java/com/kbe5/adapter/dto/request/firebase/TokenNotificationRequest.java
@@ -1,8 +1,0 @@
-package com.kbe5.adapter.dto.request.firebase;
-
-public record TokenNotificationRequest(
-        String title,
-        String content,
-        String url
-) {
-}
