--- rento-adapter/src/main/resources/application-deploy.yml
@@ -1,46 +1,0 @@
-
-logging:
-  level:
-    root: ERROR
-    com.kbe5.api: INFO
-
-spring:
-  config:
-    activate:
-      on-profile: deploy
-
-  datasource:
-    url: jdbc:mysql://${DB_HOST}:${DB_PORT}/${DB_NAME}?useSSL=false&serverTimezone=Asia/Seoul
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    username: ${DB_USER}
-    password: ${DB_PASS}
-
-  jackson:
-    time-zone: Asia/Seoul
-  jpa:
-    show-sql: true
-    hibernate:
-      ddl-auto: update
-      properties:
-        hibernate:
-          jdbc:
-            time_zone: Asia/Seoul
-  data:
-    redis:
-      host: ${REDIS_HOST}
-      port: ${REDIS_PORT}
-
-  rabbitmq:
-    host: ${RABBITMQ_HOST}
-    port: ${RABBITMQ_PORT}
-    username: ${RABBITMQ_USER_NAME}
-    password: ${RABBITMQ_PASSWORD}
-    virtual-host: /
-    ssl:
-      enabled: true
-
-fcm:
-  firebase_config_path: ${FIREBASE_PATH}
-
-server:
-  port: 8080
