--- rento-adapter/src/main/resources/application-local.yml
@@ -1,48 +1,0 @@
-spring:
-  config:
-    import: "optional:file:.env[.properties]"
-    activate:
-      on-profile: local
-
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: ${DB_URL}
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-
-  jpa:
-    database-platform: org.hibernate.dialect.MySQLDialect
-    show-sql: false
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-
-  messages:
-    basename: messages
-    encoding: UTF-8
-
-  rabbitmq:
-    username: guest # 초기값은 guest
-    password: guest # 초기값은 guest
-    host: localhost # 추후 rabbitmq 서버의 주소로 변경
-    port: 5672  # AMQP 기본 포트
-    virtual-host: / # 가상 호스트 주소
-
-  data:
-    redis:
-      host: localhost
-      port: 6379
-
-server:
-  port: 8081
-  tomcat:
-    threads:
-      max: 200
-      min-spare: 10
-    accept-count: 100
-
-fcm:
-  firebase_config_path: firebase/rento-f1d61-firebase-adminsdk-fbsvc-bd76cc388d.json
-
