--- rento-adapter/src/main/resources/application-test.yml
@@ -1,42 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: test
-
-  datasource:
-    url: jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;MODE=MYSQL
-    driver-class-name: org.h2.Driver
-    username: sa          # ← 여기 추가
-    password:              # ← 여기 추가
-
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-    hibernate:
-      ddl-auto: create
-
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-
-  rabbitmq:
-    username: guest # 초기값은 guest
-    password: guest # 초기값은 guest
-    host: localhost # 추후 rabbitmq 서버의 주소로 변경
-    port: 5672  # AMQP 기본 포트
-    virtual-host: / # 가상 호스트 주소
-
-  data:
-    redis:
-      host: localhost
-      port: 6379
-
-server:
-  tomcat:
-    threads:
-      max: 200
-      min-spare: 10
-    accept-count: 100
-
-fcm:
-  firebase_config_path: firebase/rento-f1d61-firebase-adminsdk-fbsvc-bd76cc388d.json
