--- rento-adapter/src/main/resources/application.yml
@@ -1,13 +1,0 @@
-spring:
-  config:
-    import: "optional:file:.env[.properties]"
-  profiles:
-    active: local
-  jpa:
-    show-sql: true
-    properties:
-      hibernate.jdbc.time_zone: Asia/Seoul
-
-  messages:
-    basename: messages
-    encoding: UTF-8
