+++ rento-pub/src/main/java/com/kbe5/pub/controller/EventHealthCheckController.java
@@ -1,0 +1,17 @@
+package com.kbe5.pub.controller;
+
+import io.swagger.v3.oas.annotations.Hidden;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Hidden
+@RestController
+@RequestMapping("/api/events")
+public class EventHealthCheckController {
+    @GetMapping("/health")
+    public ResponseEntity<String> health() {
+        return ResponseEntity.ok("OK");
+    }
+}
