+++ rento-pub/src/main/java/com/kbe5/pub/dto/request/firebase/TokenNotificationRequest.java
@@ -1,0 +1,8 @@
+package com.kbe5.pub.dto.request.firebase;
+
+public record TokenNotificationRequest(
+        String title,
+        String content,
+        String url
+) {
+}
